requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
boto3==1.24.75
botocore==1.27.75
cffi==1.15.1
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==3.3
jmespath==1.0.1
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.12
yarl==1.7.2
